Render IntlProvider in App.js to fix failing test
